+++ Metadata.cs
@@ -1,9 +1,9 @@
-﻿namespace GeneralImprovements
+﻿namespace PeakGeneralImprovements
 {
     internal static class Metadata
     {
-        public const string GUID = "ShaosilGaming.GeneralImprovements";
-        public const string PLUGIN_NAME = "GeneralImprovements";
-        public const string VERSION = "1.0.0";
+        public const string GUID = "ShaosilGaming.PeakGeneralImprovements";
+        public const string PLUGIN_NAME = "PeakGeneralImprovements";
+        public const string VERSION = "1.0.1";
     }
 }