+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 # Changelog
 
+### v1.0.1 - Fog rising sound fix
+* Added an option (defaults to false) to skip the pre-title screen.
+* Added an option (defaults to true) to play the fog rising sound effect on every level, not just the shore.
+* Added an option (defaults to true) to prevent the fog from having a wait timer, and only activating from height checks.
+* Added an option (defaults to true) to prevent the player from getting hungry while near campfires.
+
 ### v1.0.0 - Initial Release
+* Items can no longer fall through terrain when placed.